Integrate the Weapon and Armor classes into the battle system

+++ Assets/Scripts/Battle/BattleHUD.cs
@@ -16,7 +16,7 @@ public class BattleHUD : MonoBehaviour
 		hpSlider.value = unit.currentHP;
 	}
 
-	public void SetHP(int hp)
+	public void SetHP(float hp)
 	{
 		hpSlider.value = hp;
 	}
